refactor: simplify retry-attempt plumbing per review
- Move TUnitRetryAttemptsProperty to neutral TUnit.Engine.Reporters namespace
  (shared ToTestNode should not depend on the Html reporter namespace)
- Accept IReadOnlyList in the property and pass the list directly, dropping a
  per-retried-test array copy in ToTestNode
- Drop the redundant cast in ITestExecution.RetryAttempts (?? [])
